+++ react-components/src/components/CostsAndPricing/Duty/index.jsx
@@ -1,0 +1,14 @@
+import React from 'react'
+
+import { Input } from '@src/components/Form/Input'
+
+export const Duty = () => {
+  return (
+    <div>
+      <h2 className='h-xs p-t-l p-b-0'>Duty per unit</h2>
+      <p>To find out what duties and charges you might need to pay visit Check how to export goods and choose the commodity code that best matches your goods.</p>
+      <Input onChange={() => {}} id='asdasd' label='test' hideLabel />
+    </div>
+  )
+}
+