+++ react-components/src/components/CostsAndPricing/Overhead/index.jsx
@@ -11,7 +11,7 @@ export const Overhead = memo(({
 }) => {
   return (
     <>
-      <h2 className='h-xs p-t-l p-b-0'>Overhead costs</h2>
+      <h2 className='h-m p-v-s'>Overhead costs</h2>
       <p>These are the ongoing costs associated with running your whole business operation. These costs are important to
         know as it will help you determine how much you charge for your product in order to make a profit.</p>
       <p>Record all of your overhead costs in the table to calculate your total.</p>